fix(ci): python-full-coverage — align monitor recovery codes, line limits, and delegation tests
