+++ tests/unit/control/test_worker_parts/test_worker_part_035.py
@@ -1392,3 +1392,71 @@ class TestRunOnceStaleActiveExecutionRecoveryPart020:
         assert preserved_events == []
         assert stranded_events == []
         assert inspector.calls == []
+
+
+class TestRunOnceStaleActiveExecutionRecoveryPart019Continued:
+    @pytest.mark.unit
+    async def test_operator_refresh_before_first_preservation_preserves_live_runtime(
+        self,
+        session_factory: async_sessionmaker[AsyncSession],
+        origin_repo: Path,
+    ) -> None:
+        compose_project = "awf_refresh_before_preservation"
+        workspace_id = await _create_active_execution(
+            session_factory,
+            origin_repo,
+            "refresh-before-preservation",
+            WorkspaceStatus.pushing,
+            compose_project_name=compose_project,
+        )
+        async with session_factory() as s:
+            await WorkspaceControlService(
+                s,
+                project_stopper=_noop_project_stop,
+                cleaner_factory=_unexpected_cleaner_factory,
+            ).request_refresh_workspace(
+                workspace_id,
+                reason="operator recovery",
+                idempotency_key="refresh-before-preservation",
+            )
+            await s.commit()
+
+        inspector = _RecordingRuntimeInspector(
+            {compose_project: _live_agent_snapshot(container_id="agent-refresh-before-preserve")}
+        )
+        cleaner = _RecordingRuntimeCleaner()
+        worker = ControlWorker(
+            session_factory=session_factory,
+            provisioner=_TransitioningProvisioner(session_factory),  # type: ignore[arg-type]
+            executor=_RecordingExecutor(),
+            runtime_inspector=inspector,
+            runtime_cleaner=cleaner,
+            config=WorkerConfig(
+                poll_interval_seconds=0.01,
+                max_concurrent_executions=0,
+                stale_active_execution_scan_interval_seconds=0.0,
+            ),
+        )
+
+        assert await worker.run_once() == 0
+
+        async with session_factory() as s:
+            ws = await WorkspaceRepository(s).get(workspace_id)
+            assert ws is not None
+            preserved_events = await WorkspaceEventRepository(s).list(
+                workspace_id=workspace_id,
+                event_type=PRESERVED_EXECUTION_EVENT_TYPE,
+            )
+            stale_events = await WorkspaceEventRepository(s).list(
+                workspace_id=workspace_id,
+                event_type="workspace.stale_active_execution_detected",
+            )
+
+        assert ws.status == WorkspaceStatus.pushing.value
+        assert ws.subphase == PRESERVED_EXECUTION_SUBPHASE
+        assert ws.failure_reason is None
+        assert ws.failure_message is None
+        assert len(preserved_events) == 1
+        assert stale_events == []
+        assert cleaner.calls == []
+        assert inspector.calls == [compose_project]
